mail_id 를 우편함별 카운터에서 전역 RUID 로 바꾼다
Mail::Model 이 우편함마다 1,2,3... 을 세는 프로세스 지역 카운터로 mailId 를
배정하고 있었다. cpp-patterns.md 가 "하지 말 것" 1번으로 적어둔 바로 그
방식이고, 세 가지가 동시에 깨진다.

  - 플레이어끼리 같은 값이 나온다. DB 의 mails.mail_id 는 **단독 PK**다
  - 프로세스를 넘는 핸드오프에서 우편함을 새로 만들며 1부터 다시 시작한다
  - 로그인 적재(Seed)로 카운터를 맞춰도 그 프로세스 안에서만 유효하다

이제 Common::Ruid::Create() 로 받는다. nextMailId_ 와 그걸 뒤로 밀어두던
InsertMail/Seed 의 보정 코드가 통째로 사라졌다. AddMail 의 중복 검사는
남겼는데, 전역 유일이라 걸릴 수 없고 **걸리면 발급기가 고장 났다는 신호**라
조용히 덮어쓰지 않기 위해서다.

int64 가 되면서 와이어 포맷이 바뀌어 읽고 쓰는 쪽이 전부 딸려왔다:
Zone(Model/PlayerProcessor) -> World(PlayerManager 캐시 키, EnterZoneBody,
ToolLinkPackets) -> ProtocolClient/StressClient -> Client(C#) -> GmTool(C#).
LoginProcessor 에서 DB 의 BIGINT 를 uint32 로 좁히던 캐스트도 없앴다.

**Sql 은 손댈 것이 없었다.** mails.mail_id 가 이미 BIGINT 단독 PK 였고
주석도 "mail_id 가 RUID 라 전역 유일"이라고 적고 있었다 -- 스키마가 먼저
RUID 를 전제했고 C++ 만 뒤처져 있던 상태다.

검증: C++ 2솔루션 + Client + GmTool 빌드 에러/경고 0, GmTool 테스트 77개 통과.
ProtocolClient 로 실측 --
  mail del <없는 id> -> error=100(MailNotFound). 8바이트 파싱이 맞다는 뜻이고,
                        폭이 어긋났다면 InvalidPayload 로 떨어진다
  mail add           -> Mail Added   mailId=92976249261412353
  2초 뒤 만료 스윕    -> Mail Removed mailId=92976249261412353 (request=0)
  mail buy           -> 태스크 2건(우편 RUID + 골드 1000->900)

+++ Client/Client/Src/ClientGame.cs
@@ -671,7 +671,7 @@ public sealed class ClientGame : Game
             case MailAction.Delete:
             {
                 var writer = new BinaryPacketWriter();
-                writer.WriteUInt32(request.MailId);
+                writer.WriteInt64(request.MailId);
                 link_.Send(PacketId.C2ZMailDel, writer);
                 break;
             }
